requirements/dev.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.5
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
python-dotenv==1.0.0
django-debug-toolbar==4.2.0
